backend/requirements.txt
Django~=2.1.3
djangorestframework~=3.9.0
django-rest-swagger~=2.2.0
django-model-utils~=3.1.2
